pyproject.toml
[project]
name = "benefits"
dynamic = ["version"]
description = "Cal-ITP Benefits is an application that enables automated eligibility verification and enrollment for transit benefits onto customers’ existing contactless bank (credit/debit) cards."
readme = "README.md"
license-files = ["LICENSE"]
classifiers = ["Programming Language :: Python :: 3 :: Only"]
requires-python = ">=3.9"
maintainers = [
  { name = "Compiler LLC", email = "dev@compiler.la" }
]
dependencies = [
    "azure-keyvault-secrets==4.10.0",
    "azure-identity==1.25.1",
    "calitp-littlepay==2025.4.1",
    "Django==5.2.7",
    "django-admin-sortable2==2.3",
    "django-azure-communication-email==1.5.0",
    "django-cdt-identity @ git+https://github.com/Office-of-Digital-Services/django-cdt-identity.git@c635773a75e7f184a9382748afcef6d5a705b2d6",
    "django-csp==4.0",
    "django-google-sso==8.0.0",
    "django-multiselectfield==1.0.1",
    "eligibility-api==2025.4.1",
    "pillow==11.3.0",
    "requests==2.32.5",
    "sentry-sdk==2.48.0",
    "six==1.17.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "djlint",
    "flake8",
    "isort",
    "pre-commit",
    "setuptools_scm>=8"
]
test = [
    "coverage",
    "pytest",
    "pytest-django",
    "pytest-mock",
    "pytest-playwright", # only for writing tests. run tests using playwright Docker Compose service
    "pytest-socket",
]

[project.urls]
Changelog = "https://github.com/cal-itp/benefits/releases"
Code = "https://github.com/cal-itp/benefits"
Documentation = "https://docs.calitp.org/benefits"
Homepage = "https://www.calitp.org/#initiatives-benefits"
Issues = "https://github.com/cal-itp/benefits/issues"

[build-system]
requires = ["setuptools>=65", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 127
target-version = ['py312']
include = '\.pyi?$'

[tool.coverage.run]
branch = true
relative_files = true
omit = [
    "benefits/core/migrations/*"
]
source = ["benefits"]

[tool.djlint]
ignore = "H017,H031"
indent = 2
max_attribute_length = 127
max_line_length = 127
profile = "django"
preserve_blank_lines = true
use_gitignore = true

[tool.isort]
profile = "black"
combine_as_imports = true
line_length = 127
skip_gitignore = true

[tool.pyright]
include = ["benefits", "tests/pytest"]
typeCheckingMode = "off"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "benefits.settings"
markers = [
    "request_path: use with session_request to initialize with the given path",
]

[tool.setuptools.packages.find]
include = ["benefits*"]
namespaces = false

[tool.setuptools_scm]
# intentionally left blank, but we need the section header to activate the tool
